+++ ee/mpeg/include/libmpeg.h
@@ -51,9 +51,9 @@ typedef struct MPEGSequenceInfo {
 extern "C" {
 # endif  /* __cplusplus */
 
-void MPEG_Initialize (  int ( * ) ( void* ), void*, void* ( * ) ( void*, MPEGSequenceInfo* ), void*, long*  );
+void MPEG_Initialize (  int ( * ) ( void* ), void*, void* ( * ) ( void*, MPEGSequenceInfo* ), void*, long long*  );
 void MPEG_Destroy    ( void );
-int  ( *MPEG_Picture ) ( void*, long* );
+int  ( *MPEG_Picture ) ( void*, long long* );
 
 # ifdef __cplusplus
 }
